Application/requirements.txt
Flask==3.0.3
boto3==1.35.36
PyYAML==6.0.2
